+++ packages/exchange-consumer/src/composers/CosmosComposer.ts
@@ -1,7 +1,6 @@
 import { AccountAddress } from '@injectivelabs/ts-types'
 import { BigNumberInWei } from '@injectivelabs/utils'
 import { Coin } from '@injectivelabs/chain-api/cosmos/base/v1beta1/coin_pb'
-import { CosmosTxFee } from '@injectivelabs/exchange-api/injective_exchange_rpc_pb'
 import { MsgSendToEth } from '@injectivelabs/chain-api/injective/peggy/v1/msgs_pb'
 import snakeCaseKeys from 'snakecase-keys'
 import {
@@ -11,7 +10,6 @@ import {
 import {
   DEFAULT_BRIDGE_FEE_AMOUNT,
   DEFAULT_BRIDGE_FEE_DENOM,
-  DEFAULT_GAS_LIMIT,
 } from '../constants'
 
 export class CosmosComposer {
@@ -22,7 +20,6 @@ export class CosmosComposer {
     denom,
     bridgeFeeDenom = DEFAULT_BRIDGE_FEE_DENOM,
     bridgeFeeAmount = DEFAULT_BRIDGE_FEE_AMOUNT,
-    gasLimit = DEFAULT_GAS_LIMIT,
   }: {
     denom: string
     address: AccountAddress
@@ -30,14 +27,10 @@ export class CosmosComposer {
     amount: BigNumberInWei
     bridgeFeeDenom?: string
     bridgeFeeAmount?: string
-    gasLimit?: number
   }): Record<string, any> {
     const coinAmount = new Coin()
     coinAmount.setDenom(denom)
     coinAmount.setAmount(amount.toString())
-
-    const cosmosTxFee = new CosmosTxFee()
-    cosmosTxFee.setGas(gasLimit)
 
     const bridgeFee = new Coin()
     bridgeFee.setDenom(bridgeFeeDenom)
@@ -60,20 +53,15 @@ export class CosmosComposer {
     validatorAddress,
     amount,
     denom,
-    gasLimit = DEFAULT_GAS_LIMIT,
   }: {
     denom: string
     validatorAddress: string
     cosmosAddress: AccountAddress
     amount: BigNumberInWei
-    gasLimit?: number
   }): Record<string, any> {
     const coinAmount = new Coin()
     coinAmount.setDenom(denom)
     coinAmount.setAmount(amount.toString())
-
-    const cosmosTxFee = new CosmosTxFee()
-    cosmosTxFee.setGas(gasLimit)
 
     const cosmosMessage = new MsgDelegate()
     cosmosMessage.setAmount(coinAmount)
@@ -91,20 +79,15 @@ export class CosmosComposer {
     validatorAddress,
     amount,
     denom,
-    gasLimit = DEFAULT_GAS_LIMIT,
   }: {
     denom: string
     validatorAddress: string
     cosmosAddress: AccountAddress
     amount: BigNumberInWei
-    gasLimit?: number
   }): Record<string, any> {
     const coinAmount = new Coin()
     coinAmount.setDenom(denom)
     coinAmount.setAmount(amount.toString())
-
-    const cosmosTxFee = new CosmosTxFee()
-    cosmosTxFee.setGas(gasLimit)
 
     const cosmosMessage = new MsgUndelegate()
     cosmosMessage.setAmount(coinAmount)