+++ packages/chain-consumer/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.29](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/chain-consumer@0.0.28...@injectivelabs/chain-consumer@0.0.29) (2021-03-18)
+
+**Note:** Version bump only for package @injectivelabs/chain-consumer
+
 ## [0.0.28](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/chain-consumer@0.0.27...@injectivelabs/chain-consumer@0.0.28) (2021-03-16)
 
 ### Bug Fixes