+++ packages/networks/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.6](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/networks@0.0.5...@injectivelabs/networks@0.0.6) (2021-03-18)
+
+**Note:** Version bump only for package @injectivelabs/networks
+
 ## [0.0.5](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/networks@0.0.4...@injectivelabs/networks@0.0.5) (2021-03-14)
 
 **Note:** Version bump only for package @injectivelabs/networks